+++ cmd/searcher/main.go
@@ -148,7 +148,9 @@ func run(logger log.Logger) error {
 	service := &search.Service{
 		Store: &search.Store{
 			FetchTar: func(ctx context.Context, repo api.RepoName, commit api.CommitID) (io.ReadCloser, error) {
-				return git.Archive(ctx, repo, gitserver.ArchiveOptions{
+				// We pass in a nil checker here since searcher needs access to all data in the
+				// archive
+				return git.ArchiveReader(ctx, nil, repo, gitserver.ArchiveOptions{
 					Treeish: string(commit),
 					Format:  gitserver.ArchiveFormatTar,
 				})
@@ -158,7 +160,9 @@ func run(logger log.Logger) error {
 				for i, p := range paths {
 					pathspecs[i] = gitdomain.PathspecLiteral(p)
 				}
-				return git.Archive(ctx, repo, gitserver.ArchiveOptions{
+				// We pass in a nil checker here since searcher needs access to all data in the
+				// archive
+				return git.ArchiveReader(ctx, nil, repo, gitserver.ArchiveOptions{
 					Treeish:   string(commit),
 					Format:    gitserver.ArchiveFormatTar,
 					Pathspecs: pathspecs,
