+++ internal/gitserver/mocks_temp.go
@@ -32,9 +32,6 @@ type MockClient struct {
 	// AddrsFunc is an instance of a mock function object controlling the
 	// behavior of the method Addrs.
 	AddrsFunc *ClientAddrsFunc
-	// ArchiveFunc is an instance of a mock function object controlling the
-	// behavior of the method Archive.
-	ArchiveFunc *ClientArchiveFunc
 	// ArchiveReaderFunc is an instance of a mock function object
 	// controlling the behavior of the method ArchiveReader.
 	ArchiveReaderFunc *ClientArchiveReaderFunc
@@ -201,11 +198,6 @@ func NewMockClient() *MockClient {
 		},
 		AddrsFunc: &ClientAddrsFunc{
 			defaultHook: func() (r0 []string) {
-				return
-			},
-		},
-		ArchiveFunc: &ClientArchiveFunc{
-			defaultHook: func(context.Context, api.RepoName, ArchiveOptions) (r0 io.ReadCloser, r1 error) {
 				return
 			},
 		},
@@ -481,11 +473,6 @@ func NewStrictMockClient() *MockClient {
 				panic("unexpected invocation of MockClient.Addrs")
 			},
 		},
-		ArchiveFunc: &ClientArchiveFunc{
-			defaultHook: func(context.Context, api.RepoName, ArchiveOptions) (io.ReadCloser, error) {
-				panic("unexpected invocation of MockClient.Archive")
-			},
-		},
 		ArchiveReaderFunc: &ClientArchiveReaderFunc{
 			defaultHook: func(context.Context, authz.SubRepoPermissionChecker, api.RepoName, ArchiveOptions) (io.ReadCloser, error) {
 				panic("unexpected invocation of MockClient.ArchiveReader")
@@ -753,9 +740,6 @@ func NewMockClientFrom(i Client) *MockClient {
 		},
 		AddrsFunc: &ClientAddrsFunc{
 			defaultHook: i.Addrs,
-		},
-		ArchiveFunc: &ClientArchiveFunc{
-			defaultHook: i.Archive,
 		},
 		ArchiveReaderFunc: &ClientArchiveReaderFunc{
 			defaultHook: i.ArchiveReader,
@@ -1116,116 +1100,6 @@ func (c ClientAddrsFuncCall) Args() []interface{} {
 // invocation.
 func (c ClientAddrsFuncCall) Results() []interface{} {
 	return []interface{}{c.Result0}
-}
-
-// ClientArchiveFunc describes the behavior when the Archive method of the
-// parent MockClient instance is invoked.
-type ClientArchiveFunc struct {
-	defaultHook func(context.Context, api.RepoName, ArchiveOptions) (io.ReadCloser, error)
-	hooks       []func(context.Context, api.RepoName, ArchiveOptions) (io.ReadCloser, error)
-	history     []ClientArchiveFuncCall
-	mutex       sync.Mutex
-}
-
-// Archive delegates to the next hook function in the queue and stores the
-// parameter and result values of this invocation.
-func (m *MockClient) Archive(v0 context.Context, v1 api.RepoName, v2 ArchiveOptions) (io.ReadCloser, error) {
-	r0, r1 := m.ArchiveFunc.nextHook()(v0, v1, v2)
-	m.ArchiveFunc.appendCall(ClientArchiveFuncCall{v0, v1, v2, r0, r1})
-	return r0, r1
-}
-
-// SetDefaultHook sets function that is called when the Archive method of
-// the parent MockClient instance is invoked and the hook queue is empty.
-func (f *ClientArchiveFunc) SetDefaultHook(hook func(context.Context, api.RepoName, ArchiveOptions) (io.ReadCloser, error)) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// Archive method of the parent MockClient instance invokes the hook at the
-// front of the queue and discards it. After the queue is empty, the default
-// hook function is invoked for any future action.
-func (f *ClientArchiveFunc) PushHook(hook func(context.Context, api.RepoName, ArchiveOptions) (io.ReadCloser, error)) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *ClientArchiveFunc) SetDefaultReturn(r0 io.ReadCloser, r1 error) {
-	f.SetDefaultHook(func(context.Context, api.RepoName, ArchiveOptions) (io.ReadCloser, error) {
-		return r0, r1
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *ClientArchiveFunc) PushReturn(r0 io.ReadCloser, r1 error) {
-	f.PushHook(func(context.Context, api.RepoName, ArchiveOptions) (io.ReadCloser, error) {
-		return r0, r1
-	})
-}
-
-func (f *ClientArchiveFunc) nextHook() func(context.Context, api.RepoName, ArchiveOptions) (io.ReadCloser, error) {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *ClientArchiveFunc) appendCall(r0 ClientArchiveFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of ClientArchiveFuncCall objects describing
-// the invocations of this function.
-func (f *ClientArchiveFunc) History() []ClientArchiveFuncCall {
-	f.mutex.Lock()
-	history := make([]ClientArchiveFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// ClientArchiveFuncCall is an object that describes an invocation of method
-// Archive on an instance of MockClient.
-type ClientArchiveFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Arg1 is the value of the 2nd argument passed to this method
-	// invocation.
-	Arg1 api.RepoName
-	// Arg2 is the value of the 3rd argument passed to this method
-	// invocation.
-	Arg2 ArchiveOptions
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 io.ReadCloser
-	// Result1 is the value of the 2nd result returned from this method
-	// invocation.
-	Result1 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c ClientArchiveFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0, c.Arg1, c.Arg2}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c ClientArchiveFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0, c.Result1}
 }
 
 // ClientArchiveReaderFunc describes the behavior when the ArchiveReader
